flight-status-system/backend/requirements.txt
flask
flask-cors
kafka-python
